revert addFields fragment
