+++ packages/houdini/src/codegen/transforms/index.ts
@@ -1,6 +1,7 @@
 export { default as composeQueries } from './composeQueries'
 export { default as internalSchema } from './schema'
 export { default as list } from './list'
-export { default as addFields } from './addFields'
+export { default as typename } from './typename'
+export { default as addID } from './addID'
 export { default as paginate } from './paginate'
 export { default as fragmentVariables } from './fragmentVariables'
